refactor: rename events to notifications
Resolves #44
