--- lib/src/main/java/hu/bme/mit/ftsrg/hypernate/middleware/event/HypernateEvent.java
@@ -1,8 +1,0 @@
-/* SPDX-License-Identifier: Apache-2.0 */
-package hu.bme.mit.ftsrg.hypernate.middleware.event;
-
-import lombok.Data;
-
-/** Base class for internal Hypernate events. */
-@Data
-public abstract class HypernateEvent {}
