--- lib/src/main/java/hu/bme/mit/ftsrg/hypernate/middleware/event/TransactionBegin.java
@@ -1,10 +1,0 @@
-/* SPDX-License-Identifier: Apache-2.0 */
-package hu.bme.mit.ftsrg.hypernate.middleware.event;
-
-import lombok.EqualsAndHashCode;
-import lombok.Value;
-
-/** Event that should be fired before any transaction logic is executed. */
-@Value
-@EqualsAndHashCode(callSuper = true)
-public class TransactionBegin extends HypernateEvent {}
