+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/middleware/notification/HypernateNotification.java
@@ -1,0 +1,8 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+package hu.bme.mit.ftsrg.hypernate.middleware.notification;
+
+import lombok.Data;
+
+/** Base class for internal Hypernate notifications. */
+@Data
+public abstract class HypernateNotification {}
